issue ESA-4 fixed

+++ client/src/app/storelandingpage/store-landing-page/store-landing-page.component.ts
@@ -9,6 +9,7 @@ import { User } from 'src/app/domains/User';
 import { NewProductDialogComponent } from '../new-product-dialog/new-product-dialog.component';
 import { ProductsService } from 'src/app/products/products-page/products.service';
 import { Product } from 'src/app/domains/Product';
+import { DeleteProductDialogComponent } from '../delete-product-dialog/delete-product-dialog.component';
 
 @Component({
   selector: 'app-store-landing-page',
@@ -106,9 +107,19 @@ export class StoreLandingPageComponent implements OnInit {
   }
 
   deleteProductById(id: string){
-    this.storeService.deleteProductFromRepoById(id).subscribe(response => {
-      console.log(response);
-      this.loadProductsTable();
+    let dialogRef = this.dialog.open(DeleteProductDialogComponent, {
+      width: '600px',
+      disableClose: false
     });
+    
+    dialogRef.afterClosed().subscribe(result => {
+      if(result){
+        this.storeService.deleteProductFromRepoById(id).subscribe(response => {
+          console.log(response);
+          this.loadProductsTable();
+        });
+      }
+    });
+    
   }
 }
